Alabado sea

+++ FrontEnd/src/components/ProductView.jsx
@@ -5,10 +5,12 @@ import { Card, Badge, Row, Col, Button } from 'react-bootstrap';
 import Carousel from "react-bootstrap/Carousel";
 import {useEffect, useState} from "react";
 import Form from "react-bootstrap/Form";
-import {getController, postNoJSONController} from "../context/Actions.jsx";
+import {deleteController, getController, postController, postNoJSONController} from "../context/Actions.jsx";
 import Toast from "react-bootstrap/Toast";
+import Modal from "react-bootstrap/Modal";
 import Profile from "../assets/profile.svg";
-
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faStar,faShoppingCart, faBox, faHandHoldingUsd, faSignOut} from '@fortawesome/free-solid-svg-icons';
 
 
 const ProductView = () => {
@@ -25,41 +27,88 @@ const ProductView = () => {
     const [imageValidated, setImageValidated] = useState(false);
     const [imageList, setImageList] = useState([]);
     const [producto, setProducto] = useState({});
+    const [showCreateModal, setShowCreateModal] = useState(false);
     const [quantity, setQuantity] = useState(1);
-    const [cart, setCart] = useState([]);
+    const [rating, setRating] = useState(0);
+    const [pId, setPid] = useState(0);
+    const [comments, setComments] = useState('');
+    const [userData, setUserData] = useState({});
+    const [address,setAddress] = useState('')
+    const [email,setEmail] = useState('')
+    const [commentsP,setCommentsP] = useState([])
+
+    const [cart, setCart] = useState(() => {
+        const storedCart = localStorage.getItem('cart');
+        return storedCart ? JSON.parse(storedCart) : [];
+    });
     const navigate = useNavigate();
 
-    const addToCart = (product) => {
-        const updatedCart = [...cart, product];
-        setCart(updatedCart);
-        localStorage.setItem('cart', JSON.stringify(updatedCart));
+    
+    useEffect(() => {
+        const fetchComments = async () => {
+            try {
+                const response = await getController("/get_comments_product");
+    
+                if (!response) {
+                    setToastMessage("Fallo inesperado en la conexión");
+                    setShowToast(true);
+                }else {
+                    const body = await response.json();
+                    if (!response.ok) {
+                        setToastMessage(body.message)
+                        setShowToast(true);
+                    } else
+                        setCommentsP(body.comments);
+                        
+                }
+            } catch (error) {
+                console.log(error);
+            }
+        };
+        fetchComments()
+    }, []);
+
+    const addToCart = (product,quantity) => {
+        // Verificar si el producto ya está en el carrito
+        const existingProductIndex = cart.findIndex(item => item.id === product.id);
         console.log(cart)
-        navigate("/products")
-      };
+        if (existingProductIndex !== -1) {
+            // Si el producto ya está en el carrito, actualiza su cantidad
+            const updatedCart = cart.map((item, index) => {
+                if (index === existingProductIndex) {
+                    return { ...item, quantity: item.quantity + 1 };
+                }
+                return item;
+            });
+            setCart(updatedCart);
+            localStorage.setItem('cart', JSON.stringify(updatedCart));
+        } else {
+            // Si el producto no está en el carrito, agrégalo con cantidad 1
+            const updatedProduct = { ...product, quantity: quantity };
+            const updatedCart = [...cart, updatedProduct];
+            setCart(updatedCart);
+            localStorage.setItem('cart', JSON.stringify(updatedCart));
+        }
+        navigate("/products");
+    };
 
     useEffect(() => {
         const storedProduct = localStorage.getItem('producto');
         if (storedProduct) {
             setProducto(JSON.parse(storedProduct));
+            setPid(producto.id)
             }
     }, []);
-    
 
     useEffect(() => {
-
-        const fetchImages = async () => {
-            try {
-                const response = await getController("/get_images");
-                const body = await response.json();
-                setImageList(body.images);
-            } catch (error) {
-                console.error(error);
-            }
-        };
-
-        fetchImages()
-
+        const storedUserData = JSON.parse(localStorage.getItem('userData'));
+        if (storedUserData) {
+            setUserData(storedUserData);
+            setAddress(storedUserData.address)
+            setEmail(storedUserData.email)
+        }
     }, []);
+
 
     const handleThumbnailClick = (index) => {
         setSelectedIndex(index);
@@ -68,6 +117,49 @@ const ProductView = () => {
     const handleSelect = (selectedIndex, e) => {
         setSelectedIndex(selectedIndex);
     };
+
+   
+    const handleSubmit = async (e) =>{
+        e.preventDefault();
+
+        if (comments.trim() === '') {
+            // Mostrar un mensaje de error o realizar alguna acción
+            console.log('El campo de comentarios está vacío');
+            return; // Salir de la función si el campo está vacío
+        }
+    
+        // Recopilar los valores de los campos
+
+        let payload = {
+            email: email,
+            comment: comments,
+            rating: rating,
+            id: pId
+        }
+
+        try {
+            let response = await postController(payload, "create_comment")
+
+            if (!response) {
+                setToastMessage("Fallo inesperado en la conexión");
+                setToastBg("danger")
+                setToastTitle("Error")
+                setShowToast(true);
+            }else{
+                const body = await response.json();
+                if (response.ok)
+                    setShowCreateModal(true)
+                else{
+                    setToastBg("danger")
+                    setToastTitle("Error")
+                    setToastMessage(body.message)
+                    setShowToast(true);
+                }
+            }
+        } catch (error) {
+            console.log(error);
+        }
+    }
 
     const handleKeyDown = (e) => {
         if (e.keyCode === 38 && selectedIndex > 0) { // Flecha arriba
@@ -88,26 +180,30 @@ const ProductView = () => {
         }
       };
 
+      const handleIncrementRating = () => {
+        if (rating < 5) {
+            setRating(prevQuantity => prevQuantity + 1);
+        }
+    };
+    
+      const handleDecrementRating = () => {
+        if (rating > 0) {
+            setRating(rating - 1);
+        }
+      };
+
   return (
-    <div style={{minWidth:"1500px"}}>
+    <div style={{minWidth:"1500px", marginTop:"100px"}}>
         <Row style={{backgroundColor:"#F4F6F0", minHeight:"729px", borderRadius:"10px"}}>
             <Col>
                 <div>
                     <div className="col-md-8 py-1 px-2">
                             <div className="row bg-F4F6F0 py-3 px-2" tabIndex="0" onKeyDown={handleKeyDown}>
                                 <div className="col-auto carousel-image-button-scroll">
-                                    {imageList.map((image, index) => (
-                                        <div className="row mb-1" key={index}>
-                                            <div className="col">
-                                                    <img
-                                                        src={`data:image/png;base64,${image.image}`}
-                                                        className={`carousel-image-button ${index === selectedIndex ? 'selected' : ''}`}
-                                                        alt={image.description}
-                                                        onClick={() => handleThumbnailClick(index)}
-                                                    />
-                                            </div>
-                                        </div>
-                                    ))}
+                                <img
+                                                className="d-block w-100"
+                                                src={`data:image/png;base64,${producto.image}`} 
+                                            />
                                 </div>
                                 <div className="col flex-grow-1">
                                     <Carousel activeIndex={selectedIndex} onSelect={handleSelect}>
@@ -115,7 +211,7 @@ const ProductView = () => {
                                             <Carousel.Item key={index}>
                                                 <img
                                                     className="d-block w-100"
-                                                    src={`data:image/png;base64,${image.image}`}
+                                                    src={`data:image/png;base64,${producto.image}`}
                                                     alt={`Slide ${index}`}
                                                 />
                                             </Carousel.Item>
@@ -156,21 +252,23 @@ const ProductView = () => {
                     <div>
                         <Row style={{marginTop:"25px",width:"640px"}}>
                             <Col> 
-                            <div className="d-flex align-items-center">
-                                <button className="btn btn-outline-secondary" style={{borderColor:"#000000",borderWidth:'1px',fontSize:"20px"}} onClick={handleDecrement}>
-                                    -
-                                </button>
-                                <div style={{borderWidth:'5px',borderColor:"#000000"}}>
-                                <span className="mx-2" >{quantity}</span>
+                            <div className="d-flex align-items-center" style={{backgroundColor:"#99BA57", borderRadius:"10px",width:"150px"}}>
+                                    <button className="btn btn-outline-secondary" style={{borderColor:"#99BA57",borderWidth:'1px',fontSize:"28px",borderRadius:"10px",color:"white"}} onClick={handleDecrement}>
+                                        -
+                                    </button>
+                                    <div style={{borderWidth:'5px',borderColor:"#99BA57"}}>
+                                    <span className="mx-2 text-center" style={{fontSize:"30px",color:"white"}} >{quantity}</span>
+                                    </div>
+                                    <FontAwesomeIcon icon={faShoppingCart}  style={{ color:"white" }}/>
+                                    <button className="btn btn-outline-secondary" style={{ borderColor:"#99BA57",borderWidth:'1px',fontSize:"28px",borderRadius:"10px",color:"white"}} onClick={() => handleIncrement(producto)}>
+                                        +
+                                    </button>
+                                  
                                 </div>
-                                <button className="btn btn-outline-secondary" style={{ borderColor:"#000000",borderWidth:'1px',fontSize:"20px"}} onClick={() => handleIncrement(producto)}>
-                                    +
-                                </button>
-                            </div>
                                 </Col>
                             <Col>
                             <div style={{ display: 'flex', alignItems: 'center' }}>
-                                <button onClick={() => addToCart(producto)} className="add-to-cart-btn" style={{ height: "60px", fontSize: "28px", width: "412px", backgroundColor: "#99BA57", justifyContent: 'center', display: 'flex', alignItems: 'center' }}>Agregar al carrito</button>
+                                <button onClick={() => addToCart(producto,quantity)} className="add-to-cart-btn" style={{ height: "60px", fontSize: "28px", width: "412px", backgroundColor: "#99BA57", justifyContent: 'center', display: 'flex', alignItems: 'center' }}>Agregar al carrito</button>
                             </div>
                             </Col>
                         </Row>
@@ -196,8 +294,24 @@ const ProductView = () => {
                     </Col>
                         <Col>
                             <Row>
-                                <Col style={{ display: "flex", justifyContent: "center", alignItems: "center", height: "100px" }}>
-                                    <div>estrellas</div>
+                                <Col style={{ display: "flex", justifyContent: "center", alignItems: "center", height: "100px"}}>
+                                <div className="d-flex align-items-center" style={{backgroundColor:"#99BA57", borderRadius:"10px"}}>
+                                    <button className="btn btn-outline-secondary" style={{borderColor:"#99BA57",borderWidth:'1px',fontSize:"28px",borderRadius:"10px",color:"white"}} onClick={handleDecrementRating}>
+                                        -
+                                    </button>
+                                    <div style={{borderWidth:'5px',borderColor:"#99BA57"}}>
+                                    <span className="mx-2 text-center" style={{fontSize:"30px",color:"white"}} >{rating}</span>
+                                    </div>
+                                    <FontAwesomeIcon icon={faStar}  style={{ color:"white" }}/>
+                                    <button className="btn btn-outline-secondary" style={{ borderColor:"#99BA57",borderWidth:'1px',fontSize:"28px",borderRadius:"10px",color:"white"}} onClick={() => handleIncrementRating()}>
+                                        +
+                                    </button>
+                                </div>
+                                </Col>
+                                <Col style={{ display: "flex", justifyContent: "center", alignItems: "center", height: "100px"}}>
+                                    <button onClick={handleSubmit} className="btn" type='submit' style={{backgroundColor:"#99BA57",borderWidth:'1px',fontSize:"28px",borderRadius:"10px",color:"white",width:"200px"}}>
+                                        Subir
+                                    </button>
                                 </Col>
                             </Row>
                             <Row>
@@ -207,8 +321,8 @@ const ProductView = () => {
                                             className="form-control"
                                             rows={3}
                                             placeholder="Comenta aquí"
-                                            value={imageDescription}
-                                            onChange={(e) => {setImageDescription(e.target.value)}}
+                                            value={comments}
+                                            onChange={(e) => {setComments(e.target.value)}}
                                             maxLength={140}
                                             />
                                     </Form.Group>
@@ -222,6 +336,56 @@ const ProductView = () => {
             </Col>
         </Row>
         
+        {/*Mapeo  */}
+        {commentsP.filter(comment => comment.id === 5).map((comment, index) => (
+    
+            <Row key={index}>
+                <Col className='text-start' style={{marginTop:"20px"}}>
+                    <div style={{backgroundColor:"#F4F6F0",width:"1554px",height:"276px"}}>
+                        <Row>
+                        <Col className="text-start">
+                            <Row className="align-items-center justify-content-center" style={{ height: "100px" }}>
+                                <Col className="text-center">
+                                    <div style={{ maxWidth: "75%", maxHeight: "100%", display: "flex", justifyContent: "center", alignItems: "center" }}>
+                                        <img src={Profile} alt="" style={{ maxWidth: "75%", maxHeight: "100%"}} />
+                                    </div>
+                                </Col>
+                            </Row>
+                                <Row className="align-items-center px-5" style={{fontSize:"28px",marginTop:"100px",marginLeft:"50px"}}><Col>{comment.userName}</Col></Row>
+                        </Col>
+                            <Col>
+                                <Row>
+                                    <Col style={{ display: "flex", justifyContent: "center", alignItems: "center", height: "100px"}}>
+                                    <div className="d-flex align-items-center justify-content-center text-center" style={{backgroundColor:"#99BA57", borderRadius:"10px", width:"250px"}}>
+                                        <div style={{borderWidth:'5px', borderColor:"#99BA57"}}>
+                                            <span className="mx-2 text-center" style={{fontSize:"30px", color:"white"}} >{comment.star_rating}</span>
+                                        </div>
+                                        <FontAwesomeIcon icon={faStar}  style={{color:"white"}} />
+                                    </div> 
+                                    </Col>
+                                </Row>
+                                <Row>
+                                    <Col className="mb-1 text-start"  style={{paddingRight:"100px"}}>
+                                        <Form.Group className="mb-1"  style={{paddingRight:"100px",backgroundColor:"#F4F6F0",width:"1102px",height:"150px"}}>
+                                            <textarea
+                                                className="form-control"
+                                                rows={3}
+                                                placeholder="Comenta aquí"
+                                                disabled
+                                                value={comment.description}
+                                                maxLength={140}
+                                                />
+                                        </Form.Group>
+                                    </Col>
+                                </Row>
+                                
+                            </Col>
+                        </Row>
+                        
+                    </div>
+                </Col>
+            </Row>
+              ))}
     </div>
   )
 }
